Add files via upload
This GitHub repository is all about comprehensive project on house price prediction using linear regression. In this project, I have used the power of machine learning to develop a predictive model that estimates the selling price of residential properties based on various features and attributes.
This GitHub repository is a valuable resource for anyone interested in house price prediction using linear regression. It not only offers the codebase for the project but also provides insights, documentation, and a user-friendly guide for utilizing the predictive model.